upd all

+++ firPB.h
@@ -3,6 +3,6 @@
 
 #include "dataBuffer.h"
 
-void firPB(DataBuffer* dataBuffer, int* acRFiltre, int* acIRFiltre);
+void firPB(DataBuffer* dataBuffer, float* acRFiltre, float* acIRFiltre);
 
 #endif
